rockchip: rk3399: attempt to redirect uart2dbga to sdcard
The meat of my problem. rk3399 has the ability to redirect uart2 to
sdcard pins. This setup half works; I can push input into the uart, but
not see output.

Signed-off-by: Marty E. Plummer <hanetzer@startmail.com>

Series-to: u-boot
Series-cc: rockchip
Series-prefix: RFC
Cover-letter:
Alternative debug outputs for rockchip devices

I don't intend for the last change to be included in mainline, I just
need some form of serial output to see why archlinuxarm, alpine linux,
and Fedora workstation all fail to boot on this hardware (rk3399-gru-kevin).
Or at least, I think its failing. At the point of booting the kernel the
bottom inch or so of the screen corrupts a bit and then nothing. It
could be working, but without output on screen or uart2dbga I have no
idea what is going wrong.

I *would* use the uart-to-usb functionality of the rk3399, but as far as
I know this only works with the 2.0, and I'm not sure if those are
available on the rk3399-gru-kevin's type c ports. Do correct me if I'm
wrong.

In addition, the GRF_SOC_CON7:grf_uart_dbg_sel is, as far as I know,
undocumented, outside of the GRF_UART_DBG_SEL_C enum in grf_rk3399,
so the other two options (A and B) are just based on my assumptions.
If someone who *actually knows* could elaborate on it and correct me if
needed I'd greatly appreciate it.
END